RAGService/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
google-auth==2.40.3
google-genai==1.32.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.1
numpy==2.3.2
pgvector==0.4.1
pip==23.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
PyMuPDF==1.26.4
pypdf==6.0.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
